Fixed a problem with output from reporters becoming garbled and multiplied
Actually, I've been hunting this for 5 years...

Running cgreen with the textreporter under Emacs always duplicated and scrambled the messages. I never thought much about it until I tried to write an XML reporter, which got all its output multiplied and garbled in the output file.

This originates in the fact that at fork() any buffered output data is also copied into the childs and flushed at exit().

I thought this was a cygwin problem, but it turns out this is a long standing standard behaviour on all unixen. But it usually doesn't present itself.

Well, now that will never come back. At least not in cgreen. As long as I'm around...