user api: install and include boxed_double.h
box_double() and as_double() are needed for users
to be able to work with double constraints and assertions.

The main problem with float types is that according to
x86-64 abi they are passed via mmx registers and as the result
placed differently in memory by va_start() in case of variable
argument list, which is actively used by the implementation.
To fetch the correct argument then we should know its type,
because va_arg() of intptr_t and double will do the look up
in different memory regions. So, first of all mock api
becames not so nice, there must be indication of the argument
type.

The second problem related to the same issue is the current
implementation, which doesn't have type information during
argument fetching by va_arg.

So, the straight forward solution is to wrap doubles into
containers and to work with the pointers to the containers.
It is used internally, but must be used by library users
as well.

double mocked_double(double arg)
{
	return as_double(mock(box_double(arg)));
}

Ensure(double_works)
{
	expect(mocked_double,
	       when(arg, is_equal_to_double(3.0)),
	       will_return(box_double(4.0)));

	assert_that_double(mocked_double(3.0),
			   is_equal_to_double(4.0));
}

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>