Revised cmake package config approach.
New package configuration and version configuration files.
Hopefully much more intuitive for additions in the future.

Removed need for CMakePackageConfigHelpers by simplifying package.

Still generating config.cmake files from config.cmake.in files
so that as a developer, variables set/used in the project's
CMakelists.txt can be accessed in the config files and don't
need to be maintained / hard coded in the config files.

The only place the above feature is really used at the moment
are to generate version config using the APPLICATION_VERSION
variable.

To test the functionality I attempted to build the project with
mulitple versions of cmake. Before the changes this PR reflects
were implemented, I was unable to successfully configure the
project on a version less than the 2.8.5. The revsion to
cmake_minimum_required version reflets the result of those tests,
not a requirement for the proposed features.

-- cmake version 2.6-patch 0 --
Error on:
if (NOT(CMAKE_MAJOR_VERSION LESS 3) AND APPLE)
  if(POLICY CMP0042)
    cmake_policy(SET CMP0042 NEW) # CMake 3.0 to use @rpath on MacOSX
libraries
  endif()
endif()

Error on:
include(GNUInstallDirs)

-- cmake version 2.6-patch 4 --

Error on:
include(GNUInstallDirs)

-- cmake version 2.8.0 --

Error on:
include(GNUInstallDirs)

-- cmake version 2.8.4 --

Error on:
include(GNUInstallDirs)

-- cmake version 2.8.5 --
Success