Added documentation to README.md and revised package version logic.
README.md provides brief explanation of what's available to you
as a user when using `find_package( cgreen )`.

In the process of validating what was described in README.md,
a fix was added to the cgreen-config-verion.cmake.in.
The file name was non-standard and the versoin's "beta" indication
needed to be omitted for `find_package` compatibility.