Add tests and initial solution to 'mock()' not returning doubles
Since there is now way for cgreen to internally know that a
boxed_double is allocated area and not a number, it can't successfully
handle doubles in a number of cases. One being that
'always_expect(will_return)' with a boxed double will try to re-use a
returned memory area for subsequent returns which will likely segv.

The strategy here is to implement a somewhat type-safe handling by
introducing 'will_return_double()' and type check that from 'mock_()'
expecting to also implement a 'mock_double_()' which will do the
reverse type check.

This would also alleviate the user from the burdon of unboxing the
return value.

This is part of the radical changes to handling of double that was
triggered by #96