Add function to create markers for mock parameters that are doubles
The user will need to "mark" all double typed arguments to 'mock()'
with 'box_double()', which will also happen so that it fit into an
intptr_t. But we want to know which ones to treat as doubles to create
and compare with double semantics and type checks.

All part of work initiated by #96