Add type check for 'assert_that()' with double constraint
When a double typed constraint is applied in the context of
'assert_that()' the comparison is likely to fail. This commit add a
verification to 'assert_core_()' that a constraint of type
DOUBLE_COMPARER is not used.

N.b. the reverse check should also be added to 'assert_that_double_()'