Add index-out-of-bounds tests for vector and "Expected output"
There were no tests for the bounds checking in vector.c. Actually
there no bounds checks for '_get()' and '_remove()' only checked for
negative positions.

Since these tests, and some more, output messages to the console,
added output of "Expected output:" before that in the relevant tests
so as to signal that it is expected.