text-reporter:  Only report summary from top-level suite.
Prior to this change, if one create a top-level suite called Suite_A, and in
that suite added 2 sub-suites, subsuite_1 and subsuite_2 which had 2 tests and
4 asserts each, the output would be:

Running "Suite_A" (4 tests):
Completed "subsuite_1": 4 passes in Xms.
Completed "subsuite_2": 8 passes in Yms.
Completed "Suite_A": 8 passes in Zms.

This is misleading if you don't realize the count is cumulative.  Since the
"Running" message only occurs for the top-level suite, it makes sense for
the summary to only report for the top-level suite, and is less confusing.
The new output is:

Running "Suite_A" (4 tests):
Completed "Suite_A": 8 passes in Zms.