failure_message_for memory leak patch
Prior to this commit there was an allocated string that
was not freed after it was used. This resulted in memory
leaks that caused Valgrind tests in applications that use
libcgreen.so to fail. After this commit those tests succeed.

There are a few other leaks reported by Valgrind but this fix
remedies 202 leaks reported by the following command
`valgrind --leak-check=yes test_cgreen_c`.