Count number of tests correctly when merging many test libraries
If you use '-s <suite>' to merge multiple libraries into one,
previously cgreen-runner did not count tests in all libraries but only
the first. This fix restructures the handling of multiple libraries so
that it is possible to do that.

It does not consider any filtering of tests which reduces the number
of tests actually run.