never_expect counts as a pass when a mock is never called
Add the ability to count `never_expect` assertions as a passing test
when the mocked function is never called.
Due to current implementation a new field in `RecordedExpectation`
struct was added to ensure that a expectation that will be always
present knows if it as been triggered.

Add some tests for scenarios where `never_expect` is called when a
`expect` or `always_expect` is already present.