Add `times` constraint
In this commit the constraint that expect a mock to be invoked a specific
number of times is added.

The implementation piggyback on the `time_to_live` of a recorder expectation
to ensure it gets cleaned correctly when it is no longer expected.

To ensure a better error message a variable to store the number of times
a specific mock was called was added to the structure `RecordedExpectation`.