Add initial implementation of memory leak check
In this commit an initial implementation of the memory leak check
functionality is added.

A new constraint to do the check `check_memory_leak` is now available
Also replace all the memory management function with cgreen specific
so that we can have some more control over the memory

This patch does not widely add the memory check to all tests.
If we want this to happen we need to call the new assertion directly.

TODO:
Add the assertion to the end of the test suite.
Ensure that the memory cache is cleaned per test