Remove dependency from static library to gitrevision.h
It turned out that the problem with multiple interfering generatings
of gitrevision.h was caused by CMake not understanding that it need
not generate the same dependency twice, which it did if both the
dynamic and the static libraries depended on it. So just let the
dynamic library drive that generation.