Fix Issue #193 xml reporter broken if test failed with segmentation fault
Change logic for reporting incomplete tests to align with reporting
passing or failing tests. That is store report in temp output file
and dump it to results in xml_report_finish_test when calling
transfer_output_from method.
This way the test time is reported in the correct place on the testcase
tag