Remove singleton constraints for is_not_null and friends
This fixes #235, which was a bug when having multiple is_not_null mock
constraints in the same test. This overwrote the parameter name
causing mocks erroring out because of not finding the previous
parameters.

This was caused by re-using the singleton static constraint, so this
also affected is_null, is_true and is_false, which also have been
fixed.