Add libxml2 as a dedicated report formatter
Selected by -X/--libxml for cgreen-runner, the report formatter uses
libxml2 to construct an XML with report data. The libxml2 provides a
complete and well-supported formatting of XML and UTF-8, and is better
compatible with existing analyzing software, at the expense of build and
run-time dependency on the libxml2 library.