(build) Consistent linking to fix segfaults with CGREEN_WITH_STATIC_LIBRARY=ON
In commit 69e21d1 cmake_minimum_required was downgraded to 3.3 because
after the upgrade to 3.10 the testuite starting crashing when
CGREEN_WITH_STATIC_LIBRARY=ON. This was obviously only a workaround.
After some investigation I found that the real problem was the lack of
consistent link flag enforcement, which previously seems to have worked
only by accident. By adding the missing "-rdynamic" linker option we
can now safely upgrade to min. 3.10; the test suite passes with any
compiler (gcc-15, clang-20) and regardless of library linking.